upgrade to Hypergrid 3.0.0
